Support installing khrplatform.h again
Currently the KHR header is depended by mesa's EGL/GLES/GL headers, and
would be provided in mesa-common-dev.

So for the usage of mixing mesa's GL with mali's EGL/GLES, we have to
leave it to mesa's version and let mali uses it's own.

But for people who doesn't want mesa, let's add an option to allow
installing it in mali.

Change-Id: Id6e88e2d7f3c4f6092a3cf1f1078a6a7f284f823
Suggested-by: Gleb Mazovetskiy <glex.spb@gmail.com>
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>